Extract NodePopulator and DeserializationStatus
